Add support for changing SinkProviders at Runtime
For living updating sink providers a number of characteristics are
desirable.

We want the config to update atomically, so that it is either in one
version of the state or another but never somewhere in between. This
can be accomplished with the atom pattern, i.e. A simple proxy holding
a reference to a backing object.

Another characteristic we want is to avoid recreating all the
SinkProviders, which may contain connections to various
backends. There are two approaches to solve this either we evolve one
SinkSpecProvider to another, allowing it to carry internal
references (which are stored in immutable lists) to the next state.

The other option is for the conversion mechanism from one
SinkProviderSpec to another to be cacheable. In this case, however, we
need to eventually reap connections which are no longer needed when no
active sink refers to them. This seems like an acceptable
strategy. The caching layer can then simply be applied as a layer to
the SinkProviderFactory mechanism.